+++ changelog.md
@@ -19,6 +19,8 @@
 - Change cli to support multicluster    
 - Add ulimit docker parameter (DOCKER_ULIMIT)
 - Change DOCKER_SHM_SIZE to K3S_CONTAINER_DISK
+- Add mesos reconcile loop to periodically sync state with mesos
+
 
 ## v0.3.0